Executing SQL directly for seasonal aggregate.
Django's queryset manager does not give the view access to an iterator over
the returned rows, nor does it provide a means of accessing field keys
programatically (for key in obj.keys(): print obj[key]).

In order to avoid too-tightly-coupling the model and the view, the seasons
method has been updated to return a plain old dict of the aggregate data.

Signed-off-by: David Arbuckle <dintlu@gmail.com>