Added aggregation tables for daily and seasonal per-WBAN data to DB.
The SQL that is currently used is "fast enough", but the query optimizer
balks for a small number of WBAN stations, and computing multiple GROUP BY
operations against a 120MM row table for every request is Not Brilliant.

These tables will allow for all context data to use the ORM traditionally,
instead of relying on the aggregations manager as we do currently.

Signed-off-by: David Arbuckle <dintlu@gmail.com>