Improved template rendering time by caching stations lists.
Instead of iterating over the stations context object and generating unique
javascript for each iteration, the context object is now used to create
a JS object on the page - an array of dicts representing the data required
by the maps plugin to drive the desired functionality.

Because this asset is cached, URLs were problematic.  To avoid using an
incorrect cache key, dummy IDs are provided to the URLs in the stations
array, which are replaced in javascript as the array is iterated over.

Signed-off-by: David Arbuckle <dintlu@gmail.com>